requirements.txt
spotypy
deezer-python
